requirements.txt
python-chess
termcolor
argparse
numpy
python-chess